Moved dal out of app
