+++ src/app/index.ts
@@ -1,2 +1,2 @@
+export { type Project } from '@/dal'
 export { LINKS, ROUTES, thisRepo } from './constants'
-export { type Project } from './dal'
